+++ includes/constants.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   constants.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/08 20:41:18 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/04/17 15:06:18 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/04/18 21:53:17 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 //* This should be temporary, as "RESOLUTION * (16/9)" doesnt work sometimes.
 // # define HORIZ_RESOLUTION 1920
 //* We can change things here for reference when doing screenshots.
-# define MLX_TITLE "MiniRT Test | Shadows"
+# define MLX_TITLE "MiniRT Test | Cylinders"
 # define NONE 0
 # define LIT_ERROR "Error"