+++ src/coloring/line_collision_checker.c
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/16 19:17:25 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/30 20:43:35 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/04/18 21:52:02 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,16 +29,16 @@ int	touches_sphere(t_vectors ray, t_item sphere)
 
 int	touches_cylinder(t_vectors ray, t_item cylinder)
 {
-	/*
-TODO		No idea how to check this one yet :(
-			- Jorge
-	*/
+	float		m;
+	t_coords	p_prime;
 
-//! ~~~~ Useless stuff to make the complier happy ~~~~
-	ray.loc.x = 0;									//!
-	cylinder.type = 0;								//!
-	return (0);										//!
-//! ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+	if (!get_the_m(ray, cylinder, &m))
+		return (0);
+	p_prime = v_v_sum(ray.loc, v_f_mult(ray.dir, m));
+	if (powf(getmodule(v_v_sub(p_prime, cylinder.loc)), 2.0) <= powf(cylinder.height / 2.0, 2.0) + powf(cylinder.diameter / 2.0, 2.0))
+		return (1);
+	else
+		return (0);
 }
 	
 /*