+++ src/coloring/set_color_matrix.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   set_color_matrix.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/16 19:04:24 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/04/17 14:21:28 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/04/18 22:07:57 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ static int	set_color_pixel(t_item *items, t_item cam, t_coords *pixel)
 	t_figure_point	cp;
 
 	cp = get_closest_fig_point(dir_and_cam_2_vector(cam, *pixel), items);
-	return (rgb2int(cp.color));
+	return (/*rgb2int(cp.color)*/16755370);		//!	Remove the number and uncomment the comment at my left to reactivate the lighting/color system.
 }
 
 /*